requirements.txt
asgiref==3.11.1
attrs==26.1.0
colorama==0.4.6
Django==6.0.3
django-environ==0.13.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
factory_boy==3.3.3
Faker==40.13.0
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
mysqlclient==2.2.8
packaging==26.0
pluggy==1.6.0
Pygments==2.20.0
PyJWT==2.12.1
pytest==9.0.3
pytest-django==4.12.0
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
